Move wrapper code into helper include
This is in preparation for having several "levels" of client difficulty,
and associated wrapper code to defeat those clients.

Most of the ptrace wrapper logic we have so far can be reused for each
wrapper, and in further commits we will modify it to support providing a
handler function that implements the level-specific code for each
wrapper.